// src/webparts/calendars/models/IEventDetails.ts
export interface IEventDetails {
    gameDate: string;
    teamName: string;
    opponentTeamName: string;
    location: string;
    gameType: string;
    gameStatus: string;
    startTime: string;
    endTime: string;
    rescheduleDate: string;
    homeTeamScore: string;
    awayTeamScore: string;
    gamePrimaryKey: string;
}